chore: split new network manager tests to reduce long running / flakiness
